Renamed DataAccess.Static to DataAccess.Postgre
