+++ DataAccess/CMakeLists.txt
@@ -2,7 +2,7 @@ project(DataAccess LANGUAGES CXX)
 
 add_subdirectory(Public)
 add_subdirectory(DataAccess.Debug)
-add_subdirectory(DataAccess.Static)
+add_subdirectory(DataAccess.Postgre)
 add_subdirectory(DataAccess.SQLite)
 add_subdirectory(DataAccess.Test)
