+++ DataAccess/DataAccess.Postgre/PostgreAdapter.cpp
