+++ DataAccess/DataAccess.Postgre/UsersAmountFinder.cpp
