+++ DataAccess/DataAccess.Postgre/UsersAmountFinder.hpp
