+++ Server/Server.Core/Server.Core.Static/CMakeLists.txt
@@ -17,9 +17,9 @@ add_library(${TARGET} STATIC ${SOURCE_FILES})
 target_include_directories(${TARGET}
     INTERFACE ${CMAKE_CURRENT_SOURCE_DIR})
 target_include_directories(${TARGET} PRIVATE
-    $<TARGET_PROPERTY:DataAccess.Static,SOURCE_DIR>/..)
+    $<TARGET_PROPERTY:DataAccess.Postgre,SOURCE_DIR>/..)
 
-target_link_libraries(${TARGET} PRIVATE ${EXTRA_LIBS} CONAN_PKG::asio CONAN_PKG::yas DataAccess.Static)
+target_link_libraries(${TARGET} PRIVATE ${EXTRA_LIBS} CONAN_PKG::asio CONAN_PKG::yas DataAccess.Postgre)
 target_link_libraries(${TARGET} PUBLIC Network.Public DataAccess.Public Utility.Public)
 
 set_target_properties(${TARGET} PROPERTIES LINKER_LANGUAGE CXX)
