+++ Server/Server.Core/Server.Core.Static/Server.cpp
@@ -1,7 +1,7 @@
 #include "Server.hpp"
 #include "Network/Primitives.hpp"
-#include "DataAccess.Static/PostgreRepository.hpp"
-#include "DataAccess.Static/RepositoryUnits.hpp"
+#include "DataAccess.Postgre/PostgreRepository.hpp"
+#include "DataAccess.Postgre/RepositoryUnits.hpp"
 
 #include <future>
